Removed done callback in the close event listener of the modbusSerial in node
If the listener of a close event accepts a done callback it must be called. Else Node-red will hang, waiting for the callback to be called, while stopping flows where this node is included. In this particular case, without any asynchronous calls in the close listener, there is no need to use the done callback.